Do not add a Surface to slide background
To support more complex backgrounds - gradients, as a completely random
example - a Surface should not be used as the background for slides.
Each storyboard will need to add a Surface along with any theming, but
this makes things much easier to control as constant content could be
drawn above the Surface now. Like a gradient background. Again, as a
completely random example.