Fix clickable regions not being correctly placed
It seems that using the scale modifier or scale properties of a layout
modifier causes the clickable regions of some elements to be incorrect.
This manifests as buttons "not working" but just being somewhere else on
the screen. However, once the screen is resized, or you switch to the
overview and back, the clickable regions are correct again. Not sure if
this is a problem with how scaling was performed or if it's an actual
bug in Compose. I'm leaning towards the second. Will try and reproduce
on Android some other time.

Thankfully, this can be worked around by modifying the local density
instead and using a required size within.