Go back to FixedSize being based on placeable scaling
This helps us use intrinsic sizes in parent Composables. A handy trick
that we plan to use to make sure the overlay matches the size of the
story.

Also support intrinsics! This means that an outer intrinsic measurement
will not impact an inner layout measurement. Allowing us to continue to
use SubcomposeLayout inside a scene!