Change Highlighting to CodeStyle and remove all Composition locals
Before, to support provided font as part of the code style, everything
about code highlighting was based on composition locals. This provided
a lot of challenges, including no good way to cache results of the
highlighting. Instead, leave this up to the consumer and provide a
minimal set of classes and utilities. This allows consumers to rely
on ProvideTextStyle for the font while styling their code in a top-level
property for better caching.